settings: make global config file optional
Right now we crash in get_global_config() if the global config file does not exist.

Removing this hard requirement, positions envs and the global config file as 2 alternative credential sources as one would expect.

Relying on a global config file in the user's $HOME makes it awkward for CI, environment variables are usually prefferred especially for passing credentials.