config: use stdin instead of arguments
This changes the way `wrangler config` works, previously both the email
and the api_key were passed as arguments which would be captured by the
terminal history.

When launching the command it will prompt for an email, once stdin
receive a line it will prompt for the api_key. Both lines are values and
will be stored in the config.

Introduce a `WRANGLER_HOME` environment variable that will configure
where Wrangler defines its home directory. It will contain a `config`
directory.