refactor: WorkerBundle
This change introduce the notion of the WorkerBundle. Conceptually it
represents result of the backend (javascript, rust or webpack) computation.

The abstraction provide generalization for creating metadata and
uploading the worker to the API using a multipart form.

To maintain compatiblity with the rust backend, the WorkerBundle can be
contructed with a file in `metadata_path` property. We will use the file
directly instead of generating the metadata from the bindings.

For easier testing WorkerBundle allow to pass a `out_root` property
which will be used as base directory when persisting the bundle files.

Fixes https://github.com/cloudflare/wrangler/issues/293