Guard test against potential races
When you run `wrangler build` it will try to install the `npm`
dependencies of the user's project and, in development mode, the
one from `wranglerjs`. This can lead to races if two build were launch
at the same time. To avoid that we use a file lock and make wrangler
wait if it's already locked. However, it turns out that wrangler will
not wait and run `npm install` regardless.

This change avoids running multiple `wrangler build` at the same time
during our unit tests.