Add "bundle" type
Not sure *exactly* what this does but i basically just copied the
parts i thought were relevant from the site code. We have a src dir,
which is what needs to be watched, and an output dir, which is what
needs to be analyzed, and a build command, which is what needs to be
run when we wanna run the build.