Define Lintable, Parseable, and Validate traits
I keep making tons of changes without committing but I think that it
seems bigger than it actually is -- mostly what I've done here is
define a few traits

Lintable: a struct is able to lint itself with a given argument to check against
Parseable: a struct is able to create Self from a given input
Validate: a struct satisfies both Lintable and Parseable

The JS boilerplate is mostly done -- since we actually want to lint expressions,
and not statements, the bulk of the code at the moment is just about
getting to the expressions in each statement. Essentially a glorifed
switch case with a little recursion mixed in.

The webassembly is all todo!(), i just wanted to get typechecking working.

Basically, this is just yet another commit before we actually parse any
AST. But I did write some documentation, which is cool.