Reduce cognitive complexity in main
This replaces our massive clap App with a much simpler Cli struct that
has all the same information in an easier to understand and modify
format (types). Because everything is now typed (and because we've moved
clap::App creation out of main), we can further reduce the cognitive
complexity of main by removing lots of parameter checking and parsing,
removing that pesky clippy lint attribute.