Introduce LazyAccountId
This collects all handling of the account_id in one place, so that
errors no longer have to be handled at each call-site.

Making this a new type, rather than a function on `Manifest`, allows it
to be used for other types as well, such as `Target`.

Using a `OnceCell` rather than an `Option` has two benefits:
1. It does not allow setting the account twice, which would be a bug.
2. It allows setting the account with only an immutable reference, which
   reduces the amount of churn in calling functions.

Note that this still deserializes the empty string as not having an `account_id`
for backwards compatibility.