Don't ever try to load account ID in tests
This depends on the implicit global state of the user running the test.
Make sure it's consistent both between runs and between different users.

This also explicitly set `account_id` for tests that depend on it; previously,
the tests depended on it not mattering whether there was an account_id or not.