Remove build.rs
While working on tests for `get_deployments()`, I found that any changed to
`src/settings/toml/tests/mod.rs` would cause the whole Wrangler library to be rebuilt, not just the
tests. I found it was because Cargo reran the build.rs on every change to every file, which was far slower than necessary.
I could have changed the build script to output `rerun-if-changed`, but it turns out it wasn't even
necessary in the first place; so this improves the first-time build time as well.

- Use `CARGO_MANIFEST_DIR` instead of the working directory of the build script

  These are always the same. This also allows using the env variable directly instead of removing the quotes after the fact

- Use `--cfg debug_assertions` instead of the `PROFILE` cargo sets for build scripts

  In theory these could be different, but in practice we never build release artifacts with debug
  assertions, so this shouldn't affect the behavior unless someone explicitly builds from source and
  also modifies `Cargo.toml` to enable assertions.