Hide '{' in error message from formatting machinery
This code was doing two inconsistent things with the bail macro args:
the one in delete.rs used single `{` assuming the string would *not* get
put into std::fmt, while the one in put.rs used double `{{` assuming
that std::fmt *would* get applied to it.

This change closes the inconsistency and makes room for the bail macro
to catch issues like this statically in a future version.