Revert "Don't look for background updates unless Wrangler finished successfully"
This reverts commit 45f5ab26383787ec1ac8f17fc78c1b8f4028d7da to avoid #2161.
It's unclear to me why that would have introduced a regression,
but this whole issue is actual Undefined Behavior according to the C
standard, so it's basically impossible to predict the behavior of the program
from the source code. I unfortunately wasn't able to reproduce the regression locally so I'm not
sure whether this helps or not.