fix: expire unused assets on site upload
currently, we just delete unused assets when uploading a new version
of a workers-sites project. this can cause problems when a user
accessing an old site gets a 404 when going to a page when the new
assets haven't been pushed out to the edge yet, but the old assets have
already been deleted.

this commit addresses this by marking old assets for expiration
in five minutes, rather than deleting them outright. this should give
the new worker enough time to propogate to the edge before the old
assets are taken down.